pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-proxy"
version = "0.1.0"
description = "MCP Proxy API Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "quart>=0.19.0",
    "PyJWT>=2.10.1",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "asyncio>=3.4.3",
]

[project.scripts]
mcp-proxy = "mcp_proxy.main:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_proxy"]
